+++ VRCFaceTracking/Assets/Images/LogoIndicators/Active/Bottom.png
