+++ VRCFaceTracking/Assets/Images/LogoIndicators/Active/Top.png
