+++ VRCFaceTracking/Assets/Images/LogoIndicators/Idle/Bottom.png
