+++ VRCFaceTracking/Assets/Images/LogoIndicators/Idle/Top.png
