+++ VRCFaceTracking/Assets/Images/LogoIndicators/Uninitialized/Bottom.png
