+++ VRCFaceTracking/Assets/Images/LogoIndicators/Uninitialized/Top.png
