+++ VRCFaceTracking/Assets/Images/VRCFT.ico
