+++ VRCFaceTracking/Assets/UI/App.xaml
@@ -1,9 +1,8 @@
 ﻿<!-- Bind startup to MainStandalone.Main -->
-<Application x:Class="VRCFaceTracking.App"
+<Application x:Class="VRCFaceTracking.Assets.UI.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:VRCFaceTracking"
-             StartupUri="MainWindow.xaml"
+             StartupUri="Assets/UI/MainWindow.xaml"
              Exit="App_OnExit">
     <Application.Resources>
