+++ VRCFaceTracking/Assets/UI/MainWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="VRCFaceTracking.MainWindow"
+﻿<Window x:Class="VRCFaceTracking.Assets.UI.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
@@ -93,18 +93,14 @@
                         <RowDefinition Height="16*"/>
                         <RowDefinition Height="19*"/>
                     </Grid.RowDefinitions>
-                    <Image x:Name="VRCFTLogoTop" HorizontalAlignment="Left" Height="120" Margin="20,15,0,0" VerticalAlignment="Top" Width="120" Source="Images/LogoIndicators/Uninitialized/Top.png"/>
-                    <Image x:Name="VRCFTLogoBottom" HorizontalAlignment="Left" Height="120" Margin="20,72,0,0" VerticalAlignment="Top" Width="120" Source="Images/LogoIndicators/Uninitialized/Bottom.png"/>
+                    <Image x:Name="VRCFTLogoTop" HorizontalAlignment="Left" Height="120" Margin="20,15,0,0" VerticalAlignment="Top" Width="120" Source="../Images/LogoIndicators/Uninitialized/Top.png"/>
+                    <Image x:Name="VRCFTLogoBottom" HorizontalAlignment="Left" Height="120" Margin="20,72,0,0" VerticalAlignment="Top" Width="120" Source="../Images/LogoIndicators/Uninitialized/Bottom.png"/>
                     <TextBlock x:Name="WTVRCFT" HorizontalAlignment="Left" Margin="145,20,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Height="115" Width="185" Foreground="#FFE8E8E8" FontWeight="Bold" FontSize="20" Grid.ColumnSpan="2"><Run Text="Welcome to"/><LineBreak/><Run Text="VRCFaceTracking!"/></TextBlock>
                     <ListBox x:Name="AvatarInfo" Background="#00000000" Height="254" Margin="20,20,18,0" VerticalAlignment="Top" SelectionChanged="AvatarInfoUpdate" Grid.RowSpan="2" Grid.Column="1"/>
-                    <Button x:Name="PauseEyesButton" Content="Toggle Eye Tracking" HorizontalAlignment="Left" Height="36" Margin="190,144,0,0" VerticalAlignment="Top" Width="140" Click="PauseClickEyes" BorderThickness="1,1,1,1" />                   
-                    <Button x:Name="PauseMouthButton" Content="Toggle Lip Tracking" HorizontalAlignment="Left" Height="36" Margin="190,185,0,0" VerticalAlignment="Top" Width="140" Click="PauseClickMouth" Grid.RowSpan="2"/>
-                    <Button x:Name="ReinitializeButton" Content="Reinitialize" HorizontalAlignment="Left" Margin="190,227,0,261" Width="140" Click="ReinitializeClick" Grid.RowSpan="2"/>
-                    <TextBlock x:Name="IPText" HorizontalAlignment="Left" Margin="20,151,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Height="25" Width="32" Foreground="#FFE8E8E8" FontSize="15" Text="IP: "/>
-                    <TextBlock x:Name="PortText" HorizontalAlignment="Left" Margin="20,188,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Height="25" Width="32" Foreground="#FFE8E8E8" FontSize="15" RenderTransformOrigin="0.481,1.432" Grid.RowSpan="2"><Run Text="Port:"/><LineBreak/><Run/></TextBlock>
-                    <TextBox x:Name="IPInput" HorizontalAlignment="Left" Foreground="#FFE8E8E8" Height="26" Margin="63,149,0,0" FontSize="15" TextWrapping="Wrap" Text="127.0.0.1" VerticalAlignment="Top" Width="110" Background="#00000000" TextChanged="IPInputUpdate"/>
-                    <TextBox x:Name="PortInput" HorizontalAlignment="Left" Foreground="#FFE8E8E8" Height="26" Margin="63,188,0,0" FontSize="15" TextWrapping="Wrap" Text="9000&#xD;&#xA;" VerticalAlignment="Top" Width="110" Background="#00000000" TextChanged="PortInputUpdate" Grid.RowSpan="2"/>
-                    <ScrollViewer x:Name="Scroller" Margin="20,49,20,0" HorizontalScrollBarVisibility="Disabled" Grid.Row="1" Grid.ColumnSpan="2" Height="240" VerticalAlignment="Top">
+                    <Button x:Name="PauseEyesButton" Content="Toggle Eye Tracking" HorizontalAlignment="Left" Height="36" Margin="20,144,0,0" VerticalAlignment="Top" Width="140" Click="PauseClickEyes" BorderThickness="1,1,1,1" />                   
+                    <Button x:Name="PauseMouthButton" Content="Toggle Lip Tracking" HorizontalAlignment="Left" Height="36" Margin="20,185,0,0" VerticalAlignment="Top" Width="140" Click="PauseClickMouth" Grid.RowSpan="2"/>
+                    <Button x:Name="ReinitializeButton" Content="Reinitialize" HorizontalAlignment="Left" Margin="20,227,0,261" Width="140" Click="ReinitializeClick" Grid.RowSpan="2"/>
+                    <ScrollViewer x:Name="Scroller" Margin="20,49,20,0" HorizontalScrollBarVisibility="Auto" Grid.Row="1" Grid.ColumnSpan="2" Height="240" VerticalAlignment="Top">
                         <StackPanel Width="945" Visibility="Visible">
                             <ItemsControl ItemsSource="{Binding Source={x:Static local:Logger.ConsoleOutput}}" Width="940">
                                 <ItemsControl.ItemTemplate>
