+++ VRCFaceTracking/VRCFaceTracking.csproj
@@ -37,15 +37,15 @@
     <Prefer32bit>false</Prefer32bit>
   </PropertyGroup>
   <PropertyGroup>
-    <ApplicationIcon>VRCFT.ico</ApplicationIcon>
+    <ApplicationIcon>Assets/Images/VRCFT.ico</ApplicationIcon>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="App.xaml.cs">
+    <Compile Include="ArgsHandler.cs" />
+    <Compile Include="Assets\UI\App.xaml.cs">
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="ArgsHandler.cs" />
-    <Compile Include="MainWindow.xaml.cs">
+    <Compile Include="Assets\UI\MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
@@ -64,6 +64,7 @@
     <Compile Include="Params\Vector2.cs" />
     <Compile Include="Params\Vector3.cs" />
     <Compile Include="Params\XYParam.cs" />
+    <Compile Include="SDK\ExtTrackingModule.cs" />
     <Compile Include="VRChat.cs" />
     <Compile Include="TrackingLibs\Pimax\PimaxTracker.cs" />
     <Compile Include="TrackingLibs\Pimax\PimaxTrackingInterface.cs" />
@@ -85,6 +86,7 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
+    <EmbeddedResource Include="Assets\Images\VRCFT.ico" />
     <EmbeddedResource Include="TrackingLibs\Pimax\PimaxEyeTracker.dll" />
     <EmbeddedResource Include="TrackingLibs\SRanipal\libHTC_License.dll" />
     <EmbeddedResource Include="TrackingLibs\SRanipal\nanomsg.dll" />
@@ -145,23 +147,24 @@
     <Reference Include="WindowsFormsIntegration" />
   </ItemGroup>
   <ItemGroup>
-    <Resource Include="Images\LogoIndicators\Active\Bottom.png" />
-    <Resource Include="Images\LogoIndicators\Idle\Bottom.png" />
-    <Resource Include="Images\LogoIndicators\Idle\Top.png" />
-    <Resource Include="Images\LogoIndicators\Uninitialized\Bottom.png" />
-    <Resource Include="Images\LogoIndicators\Uninitialized\Top.png" />
-    <EmbeddedResource Include="VRCFT.ico" />
-    <Resource Include="Images\LogoIndicators\Active\Top.png" />
+    <Resource Include="Assets\Images\LogoIndicators\Active\Bottom.png" />
+    <Resource Include="Assets\Images\LogoIndicators\Active\Top.png" />
+    <Resource Include="Assets\Images\LogoIndicators\Idle\Bottom.png" />
+    <Resource Include="Assets\Images\LogoIndicators\Idle\Top.png" />
+    <Resource Include="Assets\Images\LogoIndicators\Uninitialized\Bottom.png" />
+    <Resource Include="Assets\Images\LogoIndicators\Uninitialized\Top.png" />
   </ItemGroup>
   <ItemGroup>
-    <ApplicationDefinition Include="App.xaml">
-      <Generator>MSBuild:Compile</Generator>
-      <SubType>Designer</SubType>
-    </ApplicationDefinition>
-    <Page Include="MainWindow.xaml">
+    <Page Include="Assets\UI\MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
+  </ItemGroup>
+  <ItemGroup>
+    <ApplicationDefinition Include="Assets\UI\App.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </ApplicationDefinition>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\System.Text.Json.7.0.0-preview.1.22076.8\build\System.Text.Json.targets" Condition="Exists('..\packages\System.Text.Json.7.0.0-preview.1.22076.8\build\System.Text.Json.targets')" />
